requirements.txt
django>=5.2,<6.0
Pillow>=10.0
gunicorn>=22.0
python-decouple>=3.8
dj-database-url>=2.1
psycopg2-binary>=2.9
requests>=2.31
